Silence non_upper_case_globals for log categories
For some reason r-a shows this diagnostic even though we suppress it
in Cargo.toml (I still need to create a bug report).

Making these upper case would be noisy at the call sites, and it
probably doesn't help since we don't think of these identifiers
as global variables (we never access their fields directly).
Silence the r-a warning for now.

See #12156